requirements/base.txt
Flask==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.1.3
itsdangerous==2.2.0
blinker==1.9.0
SQLAlchemy==2.0.41
Flask-SQLAlchemy==3.1.1
alembic==1.17.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.10
email-validator==2.2.0
orjson==3.10.18
argon2-cffi==23.1.0
Flask-Login==0.6.3
Flask-CORS==5.0.1
Flask-Mail==0.10.0
python-jose[cryptography]==3.3.0
Flask-Talisman==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
python-dotenv==1.1.0
cachetools==5.5.2
pydantic==2.11.5
pydantic_core==2.33.2
annotated-types==0.7.0
dependency-injector==4.48.3
typing_extensions==4.14.0
greenlet==3.2.3
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
rsa==4.9.1
sqlparse==0.5.3
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.4.0
dnspython==2.7.0
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
idna==3.10